beginnings of LINUX_NET netPath handler
git-svn-id: https://damsl.cs.indiana.edu/svn/xsp/trunk@242 6dc24a1a-a86e-4cf8-a923-33330a3f456b

+++ src/libxsp/libxsp_net_path.h
@@ -10,8 +10,8 @@ typedef struct xsp_sess_net_path_rule_crit_t {
 	struct xsp_addr src_mask;
 	struct xsp_addr dst_mask;
 
-	uint32_t src_port;
-	uint32_t dst_port;
+	uint64_t src_port;
+	uint64_t dst_port;
 
 	uint16_t l4_src_port_min;
 	uint16_t l4_src_port_max;
